16-Papyrus_Portal/requirements/base.txt
Django==4.2.6,<4.3
django-environ==0.11.2,<0.12
djangorestframework==3.14.0,<3.15
django-cors-headers==4.3.0,<4.4
django-filter==23.3,<23.4
django-autoslug==1.9.9,<2.0
django-countries==7.5.1,<7.6
django-phonenumber-field==7.2.0,<7.3
phonenumbers==8.13.23,<8.14
drf-yasg==1.21.7,<1.22
Pillow==10.1.0,<10.2
argon2-cffi==23.1.0,<23.2
pytz==2023.3,<2023.4
django-taggit==4.0.0,<4.1
redis==5.0.1,<5.1
celery==5.3.4,<5.4
flower==2.0.1,<2.1
django-celery-email==3.0.0,<3.1
django-allauth==0.58.1,<0.59
dj-rest-auth==5.0.1,<5.1
djangorestframework-simplejwt==5.3.0,<5.4
django-elasticsearch-dsl==8.0.0,<8.1
django-elasticsearch-dsl-drf==0.22.5,<0.23